+++ RELEASE_NOTES.md
@@ -15,6 +15,7 @@ Breaking changes can still occur in minor versions and patches, though, if the a
 ### ✨ New features
 
 * `UITextDocumentProxy` has a new `hasCurrentWord` property.
+* `KeyboardEnabledStateInspector` has a new `isFullAccessEnabled` enabled.